+++ test/doc_test.exs
@@ -11,5 +11,8 @@ defmodule Doc.Test do
   doctest Cldr.Number.System
   doctest Cldr.Number.Format.Compiler
   doctest Cldr.Number.Formatter.Short
-
+  doctest Cldr.Number.Formatter.Currency
+  doctest Cldr.Rbnf.Ordinal
+  doctest Cldr.Rbnf.Spellout
+  doctest Cldr.Rbnf.NumberSystem
 end